重建数据库,优化代码和标记TODO

+++ src/main/java/com/pandora/backend/repository/TokenRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.transaction.annotation.Transactional;
 import java.time.LocalDateTime;
 
-public interface TokenRepository extends JpaRepository<RefreshToken, Long> {
+public interface TokenRepository extends JpaRepository<RefreshToken, Integer> {
 
     @Transactional
     int deleteByExpiresAtBefore(LocalDateTime now);
